+++ gba_test_macros/tests/trybuild/duplicate_name.rs
@@ -1,0 +1,11 @@
+#![feature(custom_test_frameworks)]
+
+use gba_test_macros::test;
+
+#[test]
+fn foo() {}
+
+#[test]
+fn foo() {}
+
+fn main() {}