+++ gba_test/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+### Changed
+- Added `module()` to `TestCase`. `name()` should now only return the test name, not the module.
+
 ## 0.1.4 - 2024-10-09
 ### Fixed
 - Resolved issue with unstable `naked_functions` feature which caused compilation failures on the latest nightly builds.